Class of Event

Create a class of event. An event means the time point to take an action, e.g., carry out (futility, interim, final) analysis for add/remove arms, or stop a trial early.

Methods

Public methods

Event$new()

Event$get_name()

Event$get_type()

Event$get_trigger_condition()

Event$get_action()

Event$set_dry_run()

Event$execute_action()

Event$get_trigger_status()

Event$trigger_event()

Event$mute()

Event$clone()

Method new()

initialize Event

Usage

Event$new(name, type = name, trigger_condition, action = doNothing, ...)

Arguments

name

character. Name of event.

type

character vector. Event type(s) (futility, interim, final), an event can be of multiple types.

trigger_condition

function to check if this event should trigger. Return TRUE/FALSE.

action

function to execute when the event triggers.

...

arguments for trigger_condition.

Method get_name()

return name of event

Usage

Event$get_name()

Method get_type()

return type(s) of event

Usage

Event$get_type()

Method get_trigger_condition()

return trigger_condition function

Usage

Event$get_trigger_condition()

Method get_action()

return action function

Usage

Event$get_action()

Method set_dry_run()

set if dry run should be carried out for the event. For more details, refer to Controller::run.

Usage

Event$set_dry_run(dry_run)

Arguments

dry_run

logical.

Method execute_action()

execute action function

Usage

Event$execute_action(trial)

Arguments

trial

a Trial object.

Method get_trigger_status()

return trigger status

Usage

Event$get_trigger_status()

Method trigger_event()

trigger an event (always TRUE) and execute action accordingly. It calls Trial$get_data_lock_time() to lock data based on conditions implemented in Event$trigger_condition. If time that meets the condition cannot be found, Trial$get_data_lock_time() will throw an error and stop the program. This means that user needs to adjust their trigger_condition (e.g., target number of events (target_n_events) is impossible to reach).

Usage

Event$trigger_event(trial, ...)

Arguments

trial

a Trial object.

...

other arguments.

Method mute()

mute all messages (not including warnings)

Usage

Event$mute(silent)

Arguments

silent

logical.

Method clone()

The objects of this class are cloneable with this method.

Usage

Event$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

##