Class of Trial

Create a class of trial.

Methods

Public methods

Trial$new()

Trial$get_trial_data()

Trial$get_duration()

Trial$set_enroller()

Trial$get_enroller()

Trial$set_dropout()

Trial$get_dropout()

Trial$roll_back()

Trial$remove_arms()

Trial$update_sample_ratio()

Trial$add_arms()

Trial$get_name()

Trial$get_description()

Trial$get_arms()

Trial$get_arms_name()

Trial$get_number_arms()

Trial$has_arm()

Trial$get_an_arm()

Trial$get_sample_ratio()

Trial$get_number_patients()

Trial$get_number_enrolled_patients()

Trial$get_number_unenrolled_patients()

Trial$get_randomization_queue()

Trial$get_enroll_time()

Trial$enroll_a_patient()

Trial$enroll_patients()

Trial$set_current_time()

Trial$get_current_time()

Trial$get_event_tables()

Trial$get_data_lock_time_by_event_number()

Trial$get_data_lock_time_by_calendar_time()

Trial$get_locked_data()

Trial$get_locked_data_name()

Trial$get_event_number()

Trial$save_event_time()

Trial$get_event_time()

Trial$lock_data()

Trial$event_plot()

Trial$censor_trial_data()

Trial$save()

Trial$bind()

Trial$save_custom_data()

Trial$get_custom_data()

Trial$get_output()

Trial$mute()

Trial$set_tester()

Trial$get_tester()

Trial$independentIncrement()

Trial$dunnettTest()

Trial$closedTest()

Trial$get_seed()

Trial$print()

Trial$clone()

Method new()

initialize a trial

Usage

Trial$new(
  name,
  n_patients,
  duration,
  description = name,
  seed = NULL,
  enroller,
  dropout = NULL,
  ...
)

Arguments

name

character. Name of trial.

n_patients

integer. Maximum number of patients could be enrolled to the trial.

duration

Numeric. Trial duration.

description

character. Optional for description of the trial. By default it is set to be trial's name.

seed

random seed. If NULL, set.seed() will not be called, which uses seed set outside.

enroller

a function returning a vector enrollment time for patients. Its first argument is the number of enrolled patients.

dropout

a function returning a vector of dropout time for patients. Its first argument is the number of enrolled patients.

...

arguments of enroller and dropout.

Method get_trial_data()

return trial data of enrolled patients at the time of this function is called

Usage

Trial$get_trial_data()

Method get_duration()

return maximum duration of a trial

Usage

Trial$get_duration()

Method set_enroller()

set recruitment curve when initialize a trial.

Usage

Trial$set_enroller(func, ...)

Arguments

func

function to generate enrollment time. It can be built-in function like `rexp` or customized functions like `StaggeredRecruiter`.

...

arguments for func.

Method get_enroller()

get function of recruitment curve

Usage

Trial$get_enroller()

Method set_dropout()

set distribution of drop out time. This can be done when initialize a trial, or when updating a trial in adaptive design.

Usage

Trial$set_dropout(func, ...)

Arguments

func

function to generate dropout time. It can be built-in function like `rexp` or customized functions.

...

arguments for func.

Method get_dropout()

get generator of dropout time

Usage

Trial$get_dropout()

Method roll_back()

roll back data to current time of trial. By doing so, Trial$trial_data will be cut at current time, and data after then are deleted. However, Trial$enroll_time after current time are kept unchanged because that is planned enrollment curve.

Usage

Trial$roll_back()

Method remove_arms()

remove arms from a trial. enroll_patients() will be always called at the end to enroll all remaining patients after Trial$get_current_time(). This function may be used with futility analysis, dose selection, enrichment analysis (sub-population) or interim analysis (early stop for efficacy)

Usage

Trial$remove_arms(arms_name)

Arguments

arms_name

character vector. Name of arms to be removed.

Method update_sample_ratio()

update sample ratio of an arm. This could happen after an arm is added or removed. We may want to update sample ratio of unaffected arms as well. This function can only update sample ratio for one arm at a time. Once sample ratio is updated, trial data should be rolled back with updated randomization queue. Data of unenrolled patients should be re-sampled as well.

Usage

Trial$update_sample_ratio(arm_name, sample_ratio)

Arguments

arm_name

character. Name of an arm of length 1.

sample_ratio

integer. Sample ratio of the arm.

Method add_arms()

add one or more arms to the trial. enroll_patients() will be called at the end to enroll all remaining patients in private$randomization_queue. This function can be used in two scenarios. (1) add arms right after a trial is created (i.e., Trial$new(...)). sample_ratio and arms added through ... should be of same length. (2) add arms to a trial already with arm(s)

Usage

Trial$add_arms(sample_ratio, ...)

Arguments

sample_ratio

integer vector. Sample ratio for permuted block randomization. It will be appended to existing sample ratio in the trial.

...

one or more objects of class Arm. One exception in ... is an argument enforce. When enforce = TRUE, sample ratio of newly added arm. It rolls back all patients after Trial$get_current_time(), i.e. redo randomization for those patients. This can be useful to add arms one by one when creating a trial. Note that we can run Trial$add_arm(sample_ratio1, arm1) followed by Trial$add_arm(sample_ratio2, enforce = TRUE, arm2). We would expected similar result with Trial$add_arms(c(sample_ratio1, sample_ratio2), arm1, arm2). Note that these two method won't return exactly the same trial because randomization_queue were generated twice in the first approach but only once in the second approach. But statistically, they are equivalent and of the same distribution.

Method get_name()

return name of trial

Usage

Trial$get_name()

Method get_description()

return description of trial

Usage

Trial$get_description()

Method get_arms()

return a list of arms in the trial

Usage

Trial$get_arms()

Method get_arms_name()

return arms' name of trial

Usage

Trial$get_arms_name()

Method get_number_arms()

get number of arms in the trial

Usage

Trial$get_number_arms()

Method has_arm()

check if the trial has any arm. Return TRUE or FALSE.

Usage

Trial$has_arm()

Method get_an_arm()

return an arm

Usage

Trial$get_an_arm(arm_name)

Arguments

arm_name

character, name of arm to be extracted

Method get_sample_ratio()

return current sample ratio of the trial. The ratio can probably change during the trial (e.g., arm is removed or added)

Usage

Trial$get_sample_ratio(arm_names = NULL)

Arguments

arm_names

character vector of arms.

Method get_number_patients()

return number of patients when planning the trial

Usage

Trial$get_number_patients()

Method get_number_enrolled_patients()

return number of enrolled (randomized) patients

Usage

Trial$get_number_enrolled_patients()

Method get_number_unenrolled_patients()

return number of unenrolled patients

Usage

Trial$get_number_unenrolled_patients()

Method get_randomization_queue()

return randomization queue of planned but not yet enrolled patients. This function does not update randomization_queue, just return its value for debugging purpose.

Usage

Trial$get_randomization_queue(index = NULL)

Arguments

index

index to be extracted. Return all queue if NULL.

Method get_enroll_time()

return enrollment time of planned but not yet enrolled patients. This function does not update enroll_time, just return its value for debugging purpose.

Usage

Trial$get_enroll_time(index = NULL)

Arguments

index

index to extract. Return all enroll time if NULL.

Method enroll_a_patient()

assign a new patient to an arm based on planned randomization queue I may consider make this function deprecated. Instead, I'd like to have a function to sample the remaining n patients, with n as an argument. Reason: if an action (analysis, add/remove arm, early stop) is triggered based on number of event of TTE, we may need to sample ALL patients to know the accurate time point. With an enrollment function of more than one patient, we can find out the time point and roll back to then, and recover randomization queue and enrollment time for the remaining patients.

Usage

Trial$enroll_a_patient()

Method enroll_patients()

assign new patients to pre-planned randomization queue at pre-specified enrollment time.

Usage

Trial$enroll_patients(n_patients = NULL)

Arguments

n_patients

number of new patients to be enrolled. If NULL, all remaining patients in plan are enrolled. Error may be triggered if n_patients is greater than remaining patients as planned.

Method set_current_time()

set current time of a trial. Any data collected before could not be changed. private$now should be set after an event is triggered (through Event class, futility, interim, etc), an arm is added or removed as a result of an event

Usage

Trial$set_current_time(time)

Arguments

time

current calendar time of a trial.

Method get_current_time()

return current time of a trial

Usage

Trial$get_current_time()

Method get_event_tables()

count accumulative number of events (for TTE) or samples (otherwise) over calendar time (enroll time + tte for TTE, or enroll time + readout otherwise)

Usage

Trial$get_event_tables(arms = NULL)

Arguments

arms

a vector of arms' name on which the event tables are created. if NULL, all arms in the trial will be used.

Method get_data_lock_time_by_event_number()

given a set of endpoints and target number of events, determine the data lock time for Event (futility, interim, final (?)). This function does not change trial object (e.g. rolling back not yet randomized patients after the found data lock time).

Usage

Trial$get_data_lock_time_by_event_number(
  endpoints,
  arms,
  target_n_events,
  type = c("all", "any")
)

Arguments

endpoints

character vector. Data lock time is determined by a set of endpoints.

arms

a vector of arms' name on which number of events will be counted.

target_n_events

target number of events for each of the endpoints.

type

all if all target number of events are reached. any if the any target number of events is reached.

Returns

data lock time

Examples

## trial$get_data_lock_time_by_event_number(c('pfs','orr'), c(200,500), 'any')

Method get_data_lock_time_by_calendar_time()

given the calendar time to lock the data, return it with event counts of each of the endpoints.

Usage

Trial$get_data_lock_time_by_calendar_time(calendar_time, arms)

Arguments

calendar_time

numeric. Calendar time to lock the data

arms

a vector of arms' name on which number of events will be counted.

Returns

data lock time

Examples

## trial$get_data_lock_time_by_calendar_time(20)

Method get_locked_data()

return locked data for an event

Usage

Trial$get_locked_data(event_name)

Arguments

event_name

character, event name of which the locked data to be extracted.

Method get_locked_data_name()

return names of locked data

Usage

Trial$get_locked_data_name()

Method get_event_number()

return number of events at lock time

Usage

Trial$get_event_number(event_name = NULL)

Arguments

event_name

names of triggered events. Use all triggered events if NULL.

Method save_event_time()

save time of a new event.

Usage

Trial$save_event_time(event_time, event_name)

Arguments

event_time

numeric. Time of new event.

event_name

character. Name of new event.

Method get_event_time()

return event time when triggering a given event

Usage

Trial$get_event_time(event_name)

Arguments

event_name

character. Name of event.

Method lock_data()

lock data at specific calendar time. For time-to-event endpoints, their event indicator *_event should be updated accordingly. Locked data should be stored separately. DO NOT OVERWRITE/UPDATE private$trial_data! which can lose actual time-to-event information. For example, a patient may be censored at the first data lock. However, he may have event being observed in a later data lock.

Usage

Trial$lock_data(at_calendar_time, event_name)

Arguments

at_calendar_time

time point to lock trial data

event_name

assign event name as the name of locked data for future reference.

Method event_plot()

plot of cumulative number of events/samples over calendar time.

Usage

Trial$event_plot()

Method censor_trial_data()

censor trial data at calendar time

Usage

Trial$censor_trial_data(censor_at = NULL, selected_arms = NULL)

Arguments

censor_at

time of censoring. It is set to trial duration if NULL.

selected_arms

censoring is applied to selected arms (e.g., removed arms) only. If NULL, it will be set to all available arms in trial data. Otherwise, censoring is applied to user-specified arms only. This is necessary because number of events/sample size in removed arms should be fixed unchanged since corresponding event is triggered. In that case, one can update trial data by something like censor_trial_data(censor_at = event_time, selected_arms = removed_arms).

Method save()

save a single value or a one-row data frame to trial's output for further analysis/summary later.

Usage

Trial$save(value, name = "", overwrite = FALSE)

Arguments

value

value to be saved. It can be a vector (of length 1) or a data frame (of one row).

name

character to name the saved object. It will be used to name a column in trial's output if value is a vector. If value is a data frame, name will be the prefix pasted with the column name of value in trial's output. If user want to use value's column name as is in trial's output, set name to be '' as default. Otherwise, column name would be, e.g., "{name}_<{names(value)}>".

overwrite

logic. TRUE if overwriting existing entries with warning, otherwise, throwing an error and stop.

Method bind()

row bind a data frame to existing data frame. If name is not existing in Trial, then it is equivalent to Trial$save. Extra columns in value are ignored. Columns in Trial$custom_data[[name]] but not in value are filled with NA.

Usage

Trial$bind(value, name)

Arguments

value

a data frame to be saved. It can consist of one or multiple rows.

name

character. Name of object to be saved.

Method save_custom_data()

save arbitrary (number of) objects into a trial so that users can use those to control the workflow. Most common use case is to store simulation parameters to be used in action functions.

Usage

Trial$save_custom_data(value, name, overwrite = FALSE)

Arguments

value

value to be saved. Any type.

name

character. Name of the value to be accessed later.

overwrite

logic. TRUE if overwriting existing entries with warning, otherwise, throwing an error and stop.

Method get_custom_data()

return saved custom data of specified name.

Usage

Trial$get_custom_data(name)

Arguments

name

character. Name of custom data to be accessed.

Method get_output()

return a data frame of all current outputs saved by calling save.

Usage

Trial$get_output(cols = NULL, simplify = TRUE)

Arguments

cols

columns to be returned from Trial$output. If NULL, all columns are returned.

simplify

logical. Return value rather than a data frame of one column when length(col) == 1 and simplify == TRUE.

Method mute()

mute all messages (not including warnings)

Usage

Trial$mute(silent)

Arguments

silent

logical.

Method set_tester()

set a tester, now it can only be an object of class GraphicalTesting.

Usage

Trial$set_tester(tester)

Arguments

tester

object of class GraphicalTesting

Method get_tester()

return a tester, now it is a GraphicalTesting object.

Usage

Trial$get_tester()

Method independentIncrement()

calculate independent increments from a given set of events

Usage

Trial$independentIncrement(endpoint, placebo, events, planned_info, ...)

Arguments

endpoint

character. Name of time-to-event endpoint in trial's locked data.

placebo

character. String of placebo in trial's locked data.

events

a character vector of event names in the trial, e.g., listener$get_event_names().

planned_info

a vector of planned accumulative number of event of time-to-event endpoint. Note: planned_info can also be a character "oracle" so that planned number of events are set to be observed number of events, in that case inverse normal z statistics equal to one-sided logrank statistics. This is for the purpose of debugging only. In formal simulation, "oracle" should not be used if adaptation is present. Pre-fixed planned_info should be used to create weights in combination test that controls the family-wise error rate in the strong sense.

...

subset condition that is compatible with dplyr::filter. survdiff will be fitted on this subset only to compute one-sided logrank statistics. It could be useful when a trial consists of more than two arms. By default it is not specified, all data will be used to fit the model.

Returns

This function returns a data frame with columns:

p_inverse_normal

one-sided p-value for inverse normal test based on logrank test (alternative hypothesis: risk is higher in placebo arm). Accumulative data is used.

z_inverse_normal

z statistics of p_inverse_normal. Accumulative data is used.

p_lr

one-sided p-value for logrank test (alternative hypothesis: risk is higher in placebo arm). Accumulative data is used.

z_lr

z statistics of p_lr. Accumulative data is used.

info

observed accumulative event number.

planned_info

planned accumulative event number.

info_pbo

observed accumulative event number in placebo.

info_trt

observed accumulative event number in treatment arm.

wt

weights in z_inverse_normal.

Examples

\dontrun{
trial$independentIncrement('pfs', 'pbo', listener$get_event_names(), 'oracle')
}

Method dunnettTest()

carry out closed test based on Dunnett method under group sequential design.

Usage

Trial$dunnettTest(endpoint, placebo, treatments, events, planned_info, ...)

Arguments

endpoint

character of an endpoint for Dunnett test.

placebo

character. Name of placebo arm.

treatments

character vector. Name of treatment arms to be used in comparison.

events

character vector. Names of triggered events at which either adaptation is applied or statistical testing for endpoint is performed. Event in events does not need to be sorted by their triggering time.

planned_info

a data frame of planned number of events of time-to-event endpoint in each stage and each arm. Event names, i.e., events are row names of planned_info, and arm names, i.e., c(placebo, treatments) are column names. Note that it is not the accumulative but stage-wise event numbers. It is usually not easy to determine these numbers in practice, simulation may be used to get estimates. Note: planned_info can also be a character "default" so that planned_info are set to be number of newly randomized patients in the control arm in each of the stages. This assumes that event rate do not change over time and/or sample ratio between placebo and a treatment arm does not change as well, which may not be true. It is for the purpose of debugging or rapid implementation only. Using simulation to pick planned_info is recommended in formal simulation study.

...

subset condition that is compatible with dplyr::filter. survdiff will be fitted on this subset only to compute one-sided logrank statistics. It could be useful when comparison is made on a subset of treatment arms. By default it is not specified, all data (placebo plus one treatment arm at a time) in the locked data are used to fit the model.

Details

This function computes stage-wise p-values for each of the intersection hypotheses based on Dunnett test. If only one treatment arm is present, it is equivalent to compute the stage-wise p-values of elemental hypotheses. This function also computes inverse normal combination test statistics at each of the stages. The choice of planned_info can affect the calculation of stage-wise p-values. Specifically, it is used to compute the columns observed_info and p_inverse_normal in returned data frame, which will be used in Trial$closedTest(). The choice of planned_info can affect the result of Trial$closedTest() so user should chose it with caution.

Note that in Trial$closedTest(), observed_info, which is derived from planned_info, will lead to the same closed testing results up to a constant. This is because the closed test uses information fraction observed_info/sum(observed_info). As a result, setting planned_info to, e.g., 10 * planned_info should give same closed test results.

Based on numerical study, setting planned_info = "default" leads to a much higher power (roughly 10%) than setting planned_info to median of event numbers at stages, which can be determined by simulation. I am not sure if regulator would support such practice. For example, if an event (e.g., interim analysis) is triggered at a pre-specified calendar time, the number of randomized patients is random and is unknown when planning the trial. If I understand it correctly, regulator may want the information fraction in closed test (combined with Dunnett test) to be pre-fixed. In addition, this choice for planned_info assumes that the event rates does not change over time which is obviously not true. It is recommended to always use pre-fixed planned_info for restrict control of family-wise error rate. It should be pointed out that the choice of pre-fixed planned_info can affect statistical power significantly so fine-tuning may be required.

Returns

a list with element names like arm_name, arm1_name|arm2_name, arm1_name|arm2_name|arm3_name, etc., i.e., all possible combination of treatment arms in comparison. Each element is a data frame, with its column names self-explained. Specifically, the columns p_inverse_normal, observed_info, is_final can be used with GroupSequentialTest to perform significance test.

Examples

\dontrun{
trial$dunnettTest('pfs', 'pbo', c('high dose', 'low dose'), listener$get_event_names(), 'default')
}

Method closedTest()

perform closed test based on Dunnett test

Usage

Trial$closedTest(
  dunnett_test,
  treatments,
  events,
  alpha,
  alpha_spending = c("asP", "asOF")
)

Arguments

dunnett_test

object returned by Trial$dunnettTest().

treatments

character vector. Name of treatment arms to be used in comparison.

events

character vector. Names of triggered events at which significance testing for endpoint is performed in closed test. Event in events does not need to be sorted by their triggering time.

alpha

numeric. Allocated alpha.

alpha_spending

alpha spending function. It can be "asP" or "asOF". Note that theoretically it can be "asUser", but it is not tested. It may be supported in the future.

Returns

a data frame of columns arm, decision, event_at_reject, and reject_time.

Examples

\dontrun{
dt <- trial$dunnettTest(
  'pfs', 'pbo', c('high dose', 'low dose'),
  listener$get_event_names(), 'default')

trial$closedTest(dt, c('high dose', 'low dose'),
                 c('pfs interim', 'pfs final'),
                 0.025, 'asOF')
}

Method get_seed()

return random seed

Usage

Trial$get_seed()

Method print()

print a trial

Usage

Trial$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

Trial$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

risk1 <- data.frame(
  end_time = c(1, 10, 26.0, 52.0),
  piecewise_risk = c(1, 1.01, 0.381, 0.150) * exp(-3.01)
)

pfs1 <- Endpoint$new(name = 'pfs', type='tte',
          generator = PiecewiseConstantExponentialRNG,
          risk = risk1, endpoint_name = 'pfs')
orr1 <- Endpoint$new(
  name = 'orr', type = 'non-tte',
  readout = c(orr=1), generator = rbinom,
  size = 1, prob = .4)
placebo <- Arm$new(
  name = 'pbo', description = 'Placebo arm')

placebo$add_endpoints(pfs1, orr1)

risk2 <- risk1
risk2$hazard_ratio <- .8
pfs2 <- Endpoint$new(name = 'pfs', type='tte',
          generator = PiecewiseConstantExponentialRNG,
          risk = risk2, endpoint_name = 'pfs')
orr2 <- Endpoint$new(
  name = 'orr', type = 'non-tte',
  generator = rbinom, readout = c(orr=3),
  size = 1, prob = .6)
active <- Arm$new(
  name = 'ac', description = 'Active arm')

active$add_endpoints(pfs2, orr2)

## Plan a trial, Trial-3415, of up to 100 patients.
## Enrollment time follows an exponential distribution, with median 5
trial <- Trial$new(
  name = 'Trial-3415', n_patients = 100,
  seed = 31415926, duration = 100,
  enroller = rexp, rate = log(2) / 5)
trial$add_arms(sample_ratio = c(1, 2), placebo, active)
#> Arm(s) <pbo, ac> are added to the trial. 
#> Randomization is done for 100 potential patients. 
#> Data of 100 potential patients are generated for the trial with 2 arm(s) <pbo, ac>. 

## ------------------------------------------------
## Method `Trial$get_data_lock_time_by_event_number`
## ------------------------------------------------

## trial$get_data_lock_time_by_event_number(c('pfs','orr'), c(200,500), 'any')

## ------------------------------------------------
## Method `Trial$get_data_lock_time_by_calendar_time`
## ------------------------------------------------

## trial$get_data_lock_time_by_calendar_time(20)

## ------------------------------------------------
## Method `Trial$independentIncrement`
## ------------------------------------------------

if (FALSE) { # \dontrun{
trial$independentIncrement('pfs', 'pbo', listener$get_event_names(), 'oracle')
} # }

## ------------------------------------------------
## Method `Trial$dunnettTest`
## ------------------------------------------------

if (FALSE) { # \dontrun{
trial$dunnettTest('pfs', 'pbo', c('high dose', 'low dose'), listener$get_event_names(), 'default')
} # }

## ------------------------------------------------
## Method `Trial$closedTest`
## ------------------------------------------------

if (FALSE) { # \dontrun{
dt <- trial$dunnettTest(
  'pfs', 'pbo', c('high dose', 'low dose'),
  listener$get_event_names(), 'default')

trial$closedTest(dt, c('high dose', 'low dose'),
                 c('pfs interim', 'pfs final'),
                 0.025, 'asOF')
} # }