Triggering condition by calendar time

Define a condition to trigger trial milestone by calendar time. The milestone will be trigger when a trial has been running for at least the specified duration. It can be used combined with conditions specified by enrollment and eventNumber.

Usage

calendarTime(time)

Arguments

time

numeric. Calendar time to trigger a milestone of a trial.

Value

an object of class `Condition`