Class of Arm

Create a class of arm.

Methods

Public methods

Arms$new()

Arms$add_endpoints()

Arms$get_name()

Arms$get_number_endpoints()

Arms$has_endpoint()

Arms$get_endpoints()

Arms$get_endpoints_name()

Arms$generate_data()

Arms$print()

Arms$clone()

Method new()

initialize an arm

Usage

Arms$new(name, ...)

Arguments

name

name of arm, which is the arm's label in generated data

...

subset condition that is compatible with dplyr::filter. This can be used to specify inclusion criteria of an arm. By default it is not specified, i.e. all data generated by the generator specified in ....

Method add_endpoints()

add a list of endpoints to the arm

Usage

Arms$add_endpoints(...)

Arguments

...

one or more objects of class Endpoint

Method get_name()

return name of arm

Usage

Arms$get_name()

Method get_number_endpoints()

return number of endpoints in the arm

Usage

Arms$get_number_endpoints()

Method has_endpoint()

check if the arm has any endpoint. Return TRUE or FALSE.

Usage

Arms$has_endpoint()

Method get_endpoints()

return a list of endpoints in the arm

Usage

Arms$get_endpoints()

Method get_endpoints_name()

return name of endpoints registered to the arm

Usage

Arms$get_endpoints_name()

Method generate_data()

generate arm data

Usage

Arms$generate_data(n_patients_in_arm)

Arguments

n_patients_in_arm

integer. Number of patients randomized to the arm

Method print()

print an arm

Usage

Arms$print(categorical_vars = NULL)

Arguments

categorical_vars

categorical_vars character. Vector of categorical variables. This can be used to specify variables with limited distinct values as categorical variables in summary.

Method clone()

The objects of this class are cloneable with this method.

Usage

Arms$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Instead of using Arm$new, please use arm(), a user-friendly
# wrapper. See examples in ?arm