Generate enrollment time from piecewise constant uniform distribution

Accrual rate can be 10 patients/months for the first 2 months, 20 patients/ months for the second 2 months, and eventually 30 patients/months until the end of trial.

Usage

StaggeredRecruiter(n, accrual_rate)

Arguments

n

number of random numbers

accrual_rate

a data frame of columns end_time: End time for a constant rate in a time window. The start time of the first time window is 0. piecewise_rate: A constant rate in a time window. So the number of patients being recruited in that window is window length x piecewise_rate

Examples

accrual_rate <- data.frame(
  end_time = c(12, 13:17, Inf),
  piecewise_rate = c(15, 15 + 6 * (1:5), 45)
)

accrual_rate <- data.frame(
  end_time = c(3, 4, 5, 8, Inf),
  piecewise_rate = c(1, 2, 2, 3, 4)
)

StaggeredRecruiter(30, accrual_rate)
#>  [1]  1.000000  2.000000  3.000000  3.500000  4.000000  4.500000  5.000000
#>  [8]  5.333333  5.666667  6.000000  6.333333  6.666667  7.000000  7.333333
#> [15]  7.666667  8.000000  8.250000  8.500000  8.750000  9.000000  9.250000
#> [22]  9.500000  9.750000 10.000000 10.250000 10.500000 10.750000 11.000000
#> [29] 11.250000 11.500000